:recycle: s3 client simplify duplicate code DocumentStorageFileRepository
