+++ refarch-integrations/refarch-s3-integration/refarch-s3-integration-client-core/src/main/java/de/muenchen/refarch/integration/s3/client/repository/DocumentStorageFileRepository.java
@@ -3,10 +3,16 @@ package de.muenchen.refarch.integration.s3.client.repository;
 import de.muenchen.refarch.integration.s3.client.exception.DocumentStorageClientErrorException;
 import de.muenchen.refarch.integration.s3.client.exception.DocumentStorageException;
 import de.muenchen.refarch.integration.s3.client.exception.DocumentStorageServerErrorException;
+import de.muenchen.refarch.integration.s3.client.repository.presignedurl.IPresignedUrlRepository;
+import de.muenchen.refarch.integration.s3.client.repository.transfer.S3FileTransferRepository;
 import java.io.InputStream;
-import reactor.core.publisher.Mono;
+import lombok.RequiredArgsConstructor;
 
-public interface DocumentStorageFileRepository {
+@RequiredArgsConstructor
+public abstract class DocumentStorageFileRepository {
+    protected final IPresignedUrlRepository presignedUrlRepository;
+    protected final S3FileTransferRepository s3FileTransferRepository;
+
     /**
      * Gets the file specified in the parameter from the document storage.
      *
@@ -19,8 +25,11 @@ public interface DocumentStorageFileRepository {
      * @throws DocumentStorageException if the problem cannot be assigned to either the client or the S3
      *             storage or the document storage.
      */
-    byte[] getFile(String pathToFile, int expireInMinutes)
-            throws DocumentStorageException, DocumentStorageClientErrorException, DocumentStorageServerErrorException;
+    public byte[] getFile(String pathToFile, int expireInMinutes)
+            throws DocumentStorageException, DocumentStorageClientErrorException, DocumentStorageServerErrorException {
+        final String presignedUrl = this.presignedUrlRepository.getPresignedUrlGetFile(pathToFile, expireInMinutes);
+        return this.s3FileTransferRepository.getFile(presignedUrl);
+    }
 
     /**
      * Retrieves the file size of a file specified in the parameter from the document storage.
@@ -33,7 +42,7 @@ public interface DocumentStorageFileRepository {
      * @throws DocumentStorageException if the problem cannot be assigned to either the client or the S3
      *             storage or the document storage.
      */
-    Mono<Long> getFileSize(String pathToFile)
+    abstract Long getFileSize(String pathToFile)
             throws DocumentStorageClientErrorException, DocumentStorageServerErrorException, DocumentStorageException;
 
     /**
@@ -48,8 +57,11 @@ public interface DocumentStorageFileRepository {
      * @throws DocumentStorageException if the problem cannot be assigned to either the client or the S3
      *             storage or the document storage.
      */
-    InputStream getFileInputStream(String pathToFile, int expireInMinutes)
-            throws DocumentStorageException, DocumentStorageClientErrorException, DocumentStorageServerErrorException;
+    public InputStream getFileInputStream(final String pathToFile, final int expireInMinutes)
+            throws DocumentStorageException, DocumentStorageClientErrorException, DocumentStorageServerErrorException {
+        final String presignedUrl = this.presignedUrlRepository.getPresignedUrlGetFile(pathToFile, expireInMinutes);
+        return this.s3FileTransferRepository.getFileInputStream(presignedUrl);
+    }
 
     /**
      * Saves the file specified in the parameter to the document storage.
@@ -63,8 +75,11 @@ public interface DocumentStorageFileRepository {
      * @throws DocumentStorageException if the problem cannot be assigned to either the client or the S3
      *             storage or the document storage.
      */
-    void saveFile(String pathToFile, byte[] file, int expireInMinutes)
-            throws DocumentStorageException, DocumentStorageClientErrorException, DocumentStorageServerErrorException;
+    public void saveFile(final String pathToFile, final byte[] file, final int expireInMinutes)
+            throws DocumentStorageException, DocumentStorageClientErrorException, DocumentStorageServerErrorException {
+        final String presignedUrl = this.presignedUrlRepository.getPresignedUrlSaveFile(pathToFile, expireInMinutes);
+        this.s3FileTransferRepository.saveFile(presignedUrl, file);
+    }
 
     /**
      * Saves the file specified in the parameter to the document storage.
@@ -78,8 +93,11 @@ public interface DocumentStorageFileRepository {
      * @throws DocumentStorageException if the problem cannot be assigned to either the client or the S3
      *             storage or the document storage.
      */
-    void saveFileInputStream(String pathToFile, InputStream file, int expireInMinutes)
-            throws DocumentStorageException, DocumentStorageClientErrorException, DocumentStorageServerErrorException;
+    public void saveFileInputStream(final String pathToFile, final InputStream file, final int expireInMinutes)
+            throws DocumentStorageException, DocumentStorageClientErrorException, DocumentStorageServerErrorException {
+        final String presignedUrl = this.presignedUrlRepository.getPresignedUrlSaveFile(pathToFile, expireInMinutes);
+        this.s3FileTransferRepository.saveFileInputStream(presignedUrl, file);
+    }
 
     /**
      * Updates the file specified in the parameter to the document storage.
@@ -93,8 +111,11 @@ public interface DocumentStorageFileRepository {
      * @throws DocumentStorageException if the problem cannot be assigned to either the client or the S3
      *             storage or the document storage.
      */
-    void updateFile(String pathToFile, byte[] file, int expireInMinutes)
-            throws DocumentStorageException, DocumentStorageClientErrorException, DocumentStorageServerErrorException;
+    public void updateFile(final String pathToFile, final byte[] file, final int expireInMinutes)
+            throws DocumentStorageException, DocumentStorageClientErrorException, DocumentStorageServerErrorException {
+        final String presignedUrl = this.presignedUrlRepository.getPresignedUrlUpdateFile(pathToFile, expireInMinutes);
+        this.s3FileTransferRepository.updateFile(presignedUrl, file);
+    }
 
     /**
      * Updates the file specified in the parameter withinq the document storage.
@@ -108,8 +129,11 @@ public interface DocumentStorageFileRepository {
      * @throws DocumentStorageException if the problem cannot be assigned to either the client or the S3
      *             storage or the document storage.
      */
-    void updateFileInputStream(String pathToFile, InputStream file, int expireInMinutes)
-            throws DocumentStorageException, DocumentStorageClientErrorException, DocumentStorageServerErrorException;
+    public void updateFileInputStream(final String pathToFile, final InputStream file, final int expireInMinutes)
+            throws DocumentStorageException, DocumentStorageClientErrorException, DocumentStorageServerErrorException {
+        final String presignedUrl = this.presignedUrlRepository.getPresignedUrlUpdateFile(pathToFile, expireInMinutes);
+        this.s3FileTransferRepository.updateFileInputStream(presignedUrl, file);
+    }
 
     /**
      * Deletes the file specified in the parameter from the document storage.
@@ -122,6 +146,9 @@ public interface DocumentStorageFileRepository {
      * @throws DocumentStorageException if the problem cannot be assigned to either the client or the S3
      *             storage or the document storage.
      */
-    void deleteFile(String pathToFile, int expireInMinutes)
-            throws DocumentStorageException, DocumentStorageClientErrorException, DocumentStorageServerErrorException;
+    public void deleteFile(final String pathToFile, final int expireInMinutes)
+            throws DocumentStorageException, DocumentStorageClientErrorException, DocumentStorageServerErrorException {
+        final String presignedUrl = this.presignedUrlRepository.getPresignedUrlDeleteFile(pathToFile, expireInMinutes);
+        this.s3FileTransferRepository.deleteFile(presignedUrl);
+    }
 }
